moved the trajectory constants to SwerveConstants
