+++ src/main/java/frc/robot/Constants.java
@@ -1,8 +1,6 @@
 package frc.robot;
 
 import edu.wpi.first.math.geometry.Rotation2d;
-import edu.wpi.first.math.trajectory.TrapezoidProfile;
-import edu.wpi.first.math.trajectory.TrapezoidProfile.Constraints;
 import edu.wpi.first.math.util.Units;
 import frc.robot.subsystems.swerve.SwerveConstants.*;
 
@@ -88,47 +86,6 @@ public final class Constants {
 
     public static final double BLUE_PASSING_X = 1.343673825263977;
     public static final double BLUE_PASSING_Y = 6.969234943389893;
-  }
-
-  public static final class TrajectoryConstants {
-
-    public static final double DRIVE_BASE_DIAMETER =
-        Math.sqrt(Math.pow(DriveConstants.TRACK_WIDTH, 2) + Math.pow(DriveConstants.WHEEL_BASE, 2));
-    public static final double MAX_SPEED = 5.0;
-    public static final double MAX_ACCELERATION = 3;
-
-    public static final double AUTO_TRANSLATION_P = 1.5; // 1.7
-    public static final double AUTO_TRANSLATION_D = 0.2;
-    public static final double AUTO_THETA_P = 4.5; // 5
-    public static final double AUTO_THETA_D = 0.4;
-
-    public static final double AUTO_SHOOT_HEADING_OFFSET = 2;
-
-    public static final double MAX_ANGULAR_SPEED_RADIANS_PER_SECOND = 2;
-    public static final double MAX_ANGULAR_SPEED_RADIANS_PER_SECOND_SQUARED = 2;
-
-    // Constraint for the motion profiled robot angle controller
-    public static final TrapezoidProfile.Constraints THETA_CONTROLLER_CONSTRAINTS =
-        new TrapezoidProfile.Constraints(
-            MAX_ANGULAR_SPEED_RADIANS_PER_SECOND, MAX_ANGULAR_SPEED_RADIANS_PER_SECOND_SQUARED);
-
-    public static final double X_TOLERANCE = 0.02;
-    public static final double Y_TOLERANCE = 0.02;
-    public static final double THETA_TOLERANCE = 1.25;
-
-    // Note Detection Driving Constants
-    public static final double AUTO_ALIGN_TRANSLATIONAL_P = 3;
-    public static final double AUTO_ALIGN_TRANSLATIONAL_I = 0;
-    public static final double AUTO_ALIGN_TRANSLATIONAL_D = 0;
-
-    public static Constraints AUTO_ALIGN_TRANSLATION_CONSTRAINTS = new Constraints(5, 2);
-
-    public static final double AUTO_ALIGN_ROTATIONAL_P = 3;
-    public static final double AUTO_ALIGN_ROTATIONAL_I = 0;
-    public static final double AUTO_ALIGN_ROTATIONAL_D = 0;
-
-    public static Constraints AUTO_ALIGN_ROTATIONAL_CONSTRAINTS =
-        new Constraints(DriveConstants.MAX_ANGULAR_SPEED_RADIANS_PER_SECOND, 2);
   }
 
   public static final class JoystickConstants {
